feat: fetch templates with fallback
