+++ src/utilities/fetchAndExtractZip.utils.ts
@@ -4,15 +4,7 @@ import i18n from "i18next";
 import JSZip from "jszip";
 import { memoize } from "lodash";
 
-import {
-	DirectoryNode,
-	FileNode,
-	FileStructure,
-	FileWithContent,
-	MarkdownAttributes,
-	ProcessedZipOutput,
-} from "@interfaces/utilities";
-import { remoteTemplatesRepositoryURL } from "@src/constants";
+import { DirectoryNode, FileNode, FileStructure, MarkdownAttributes, ProcessedZipOutput } from "@interfaces/utilities";
 import { RemoteTemplateCardType, RemoteTemplateCategory } from "@src/types/components";
 
 const isFileNode = memoize((node: FileNode | DirectoryNode): node is FileNode => node?.type === "file");
@@ -62,13 +54,57 @@ const processZipContent = async (zip: JSZip): Promise<FileStructure> => {
 	return fileStructure;
 };
 
+const fetchZipFromUrl = async (url: string): Promise<ArrayBuffer> => {
+	const { data: zipData } = await axios.get(url, {
+		responseType: "arraybuffer",
+		timeout: 30000,
+		validateStatus: (status) => status === 200,
+	});
+
+	return zipData;
+};
+
 export const fetchAndUnpackZip = async (): Promise<ProcessedZipOutput> => {
+	const githubUrl = "https://raw.githubuserconte---nt.com/autokitteh/kittehub/refs/heads/release/dist.zip";
+	const fallbackUrl = "/assets/templates/kittehub.zip";
+
 	try {
-		const { data: zipData } = await axios.get(remoteTemplatesRepositoryURL, {
-			responseType: "arraybuffer",
-			timeout: 30000,
-			validateStatus: (status) => status === 200,
-		});
+		let zipData: ArrayBuffer;
+
+		try {
+			zipData = await fetchZipFromUrl(githubUrl);
+			// eslint-disable-next-line no-console
+			console.log(
+				i18n.t("fetchAndExtract.fetchedFromGithub", {
+					namespace: "utilities",
+				})
+			);
+		} catch (githubError) {
+			console.warn(
+				i18n.t("fetchAndExtract.githubFetchFailed", {
+					namespace: "utilities",
+					error: githubError instanceof Error ? githubError.message : "Unknown error",
+				})
+			);
+
+			try {
+				zipData = await fetchZipFromUrl(fallbackUrl);
+				// eslint-disable-next-line no-console
+				console.log(
+					i18n.t("fetchAndExtract.fetchedFromFallback", {
+						namespace: "utilities",
+					})
+				);
+			} catch (fallbackError) {
+				throw new Error(
+					i18n.t("fetchAndExtract.allSourcesFailed", {
+						namespace: "utilities",
+						githubError: githubError instanceof Error ? githubError.message : "Unknown error",
+						fallbackError: fallbackError instanceof Error ? fallbackError.message : "Unknown error",
+					})
+				);
+			}
+		}
 
 		const zip = new JSZip();
 		const content = await zip.loadAsync(zipData);
@@ -114,6 +150,11 @@ const getDirectoryStructure = (fileStructure: FileStructure, targetPath: string)
 
 	return currentLevel;
 };
+
+interface FileWithContent {
+	path: string;
+	content: string;
+}
 
 const getAllFilesInDirectory = (structure: FileStructure, currentPath: string = ""): FileWithContent[] => {
 	const files: FileWithContent[] = [];
